feat: update template and add metrics (#20)
I'm trying to keep the template very simple (not assuming service, it
could be an event handler) but we always want metrics, so providing an
opinionated basic set of metrics to help (but of course people can
change this and do it differently).

Also using `go tool` instead of having versions on Makefile.
Also removing golangci lint, linting is great and I would like to add
something back soon but I really can't take golangci lint anymore...that
shit is bloated/slow/annoying and with pretty bad defaults and then you
end up with big configurations...just not happy with it for a long time.